Added 2 new JUnit files for rating and song classes
EDITS:

Song class: Added a new method -> isFavorite that returns a boolean if song is favorited.

Rating: Added a new  method isEquals, (cant override a method in a enum).

testAll: Added the new JUnit test classes to class, also changed the method to be a Junit @Test method, instead of being in the main()

~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Comments: Both the parseFile and Song class currently have 100% coverage, the Rating class currently has 98% coverage, it is missing a branch. I left more details commented in the rating class. (I say currently incase if we add more methods to the classes).