Added libraryGenreSearch and Coverage/tests
Added the functionality of being able to search for a genre is the library search and add a song to a playlist based on the genre search.

Added getSongsByGenre in the LibraryModel Class to get a shallow copy of all the songs that are in that genre.

Added getGenreList in the LibraryModel Class to get a List of all the genres in the library. Represented as strings, so there isn't any escaping references.

Added necessary Test cases and classes to test all methods in model and to get branch coverage to 99.8%